+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - com.dwarfeng.subgrade.impl.service.AbstractBatchCrudService。
   - com.dwarfeng.subgrade.impl.service.AbstractCrudService。
   - com.dwarfeng.subgrade.impl.service.AbstractEntireLookupService。
+  - com.dwarfeng.subgrade.impl.service.AbstractPresetLookupService。
 
 - 优化部分服务的实现的继承关系。
   - com.dwarfeng.subgrade.impl.service.CustomBatchCrudService。
@@ -18,6 +19,7 @@
   - com.dwarfeng.subgrade.impl.service.GeneralCrudService。
   - com.dwarfeng.subgrade.impl.service.GeneralEntireLookupService。
   - com.dwarfeng.subgrade.impl.service.DaoOnlyEntireLookupService。
+  - com.dwarfeng.subgrade.impl.service.DaoOnlyPresetLookupService。
 
 - 优化部分接口及其实现的参数名称。
   - com.dwarfeng.subgrade.stack.service.BatchCrudService。